final code finalized

+++ trunk/ece477/final/atod.h
@@ -3,4 +3,4 @@
  * and making A-D readings, it will return the A-D result as an unsigned int
  */
 
-int adconvert(int adcx);
+uint16_t adconvert(int adcx);
